functions/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudevents==1.9.0
cryptography==41.0.2
deprecation==2.1.0
firebase-admin==6.2.0
firebase-functions==0.1.1
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.4.0
functions-framework==3.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-events==0.10.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
openai==0.27.8
packaging==23.1
pinecone-client==2.2.4
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.8.0
pyparsing==3.1.0
pyppeteer==1.0.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scrapingbee==1.2.0
six==1.16.0
soupsieve==2.4.1
torch==2.0.1
transformers==4.31.0
tqdm==4.65.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
watchdog==3.0.0
websockets==10.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
